+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ This document contains the FSL-MRS release history in reverse chronological orde
 2.1.5 (Monday 25th April 2023)
 ------------------------------
 - Improved speed of coil combination for MRSI
-- Noise or covariance matrix may be supplied directly for coil combination (`fsl_mrs_proc`).
+- Noise or covariance matrix may be supplied directly for pre-whitening in coil combination (`fsl_mrs_proc`, `fsl_mrs_preproc`, `fsl_mrs_preproc_edit`).
 - More robust tests for coil combination.
 - Fix 1D projection method for simulation of edited sequences